feat: 도메인 별 패키지 구조 변경
- 도메인 별 패키지 구조 변경
- package -> item
- packageoption -> itemoption
