--- src/main/java/com/moa/moa/api/address/domain/entity/Address.java
@@ -1,10 +1,0 @@
-package com.moa.moa.api.address.domain.entity;
-
-import org.locationtech.jts.geom.Point;
-
-public class Address {
-    private String address;
-    private String addressDetail;
-    private Point location;
-    private String url;
-}
