--- src/main/java/com/moa/moa/api/amenity/domain/entity/Amenity.java
@@ -1,5 +1,0 @@
-package com.moa.moa.api.amenity.domain.entity;
-
-public class Amenity {
-    private String name;
-}
