+++ src/main/java/com/moa/moa/api/category/category/domain/entity/Category.java
@@ -1,0 +1,5 @@
+package com.moa.moa.api.category.category.domain.entity;
+
+public class Category {
+    private String name;
+}
