--- src/main/java/com/moa/moa/api/category/domain/entity/Category.java
@@ -1,5 +1,0 @@
-package com.moa.moa.api.category.domain.entity;
-
-public class Category {
-    private String name;
-}
