+++ src/main/java/com/moa/moa/api/cs/question/domain/entity/Question.java
@@ -1,0 +1,11 @@
+package com.moa.moa.api.cs.question.domain.entity;
+
+import com.moa.moa.api.member.member.domain.entity.Member;
+
+public class Question {
+    private Member member;
+    private String type;
+    private String title;
+    private String content;
+    private String status;
+}
