--- src/main/java/com/moa/moa/api/custom/domain/entity/Custom.java
@@ -1,12 +1,0 @@
-package com.moa.moa.api.custom.domain.entity;
-
-import com.moa.moa.api.member.domain.entity.Member;
-
-public class Custom {
-    private Member member;
-    private String gender;
-    private String nickname;
-    private String packageType;
-    private String clothesType;
-    private String equipmentType;
-}
