--- src/main/java/com/moa/moa/api/member/domain/entity/Member.java
@@ -1,7 +1,0 @@
-package com.moa.moa.api.member.domain.entity;
-
-public class Member {
-    private String email;
-    private String nickname;
-    private String role;
-}
