+++ src/main/java/com/moa/moa/api/member/member/domain/entity/Member.java
@@ -1,0 +1,7 @@
+package com.moa.moa.api.member.member.domain.entity;
+
+public class Member {
+    private String email;
+    private String nickname;
+    private String role;
+}
