--- src/main/java/com/moa/moa/api/naverreview/domain/entity/NaverReview.java
@@ -1,9 +1,0 @@
-package com.moa.moa.api.naverreview.domain.entity;
-
-import com.moa.moa.api.shop.domain.entity.Shop;
-
-public class NaverReview {
-    private Shop shop;
-    private Double avgScore;
-    private Long totalReview;
-}
