+++ src/main/java/com/moa/moa/api/place/amenity/domain/entity/Amenity.java
@@ -1,0 +1,5 @@
+package com.moa.moa.api.place.amenity.domain.entity;
+
+public class Amenity {
+    private String name;
+}
