+++ src/main/java/com/moa/moa/api/place/placeamenity/domain/entity/PlaceAmenity.java
@@ -1,0 +1,10 @@
+package com.moa.moa.api.place.placeamenity.domain.entity;
+
+import com.moa.moa.api.place.amenity.domain.entity.Amenity;
+import com.moa.moa.api.place.place.domain.entity.Place;
+
+public class PlaceAmenity {
+    private Place place;
+    private Amenity amenity;
+    private Boolean use;
+}
