+++ src/main/java/com/moa/moa/api/place/slope/domain/entity/Slope.java
@@ -1,0 +1,9 @@
+package com.moa.moa.api.place.slope.domain.entity;
+
+import com.moa.moa.api.place.place.domain.entity.Place;
+
+public class Slope {
+    private Place place;
+    private String name;
+    private String level;
+}
