--- src/main/java/com/moa/moa/api/placeamenity/domain/entity/PlaceAmenity.java
@@ -1,10 +1,0 @@
-package com.moa.moa.api.placeamenity.domain.entity;
-
-import com.moa.moa.api.amenity.domain.entity.Amenity;
-import com.moa.moa.api.place.domain.entity.Place;
-
-public class PlaceAmenity {
-    private Place place;
-    private Amenity amenity;
-    private Boolean use;
-}
