--- src/main/java/com/moa/moa/api/placeshop/domain/entity/placeShop.java
@@ -1,9 +1,0 @@
-package com.moa.moa.api.placeshop.domain.entity;
-
-import com.moa.moa.api.place.domain.entity.Place;
-import com.moa.moa.api.shop.domain.entity.Shop;
-
-public class placeShop {
-    private Place place;
-    private Shop shop;
-}
