--- src/main/java/com/moa/moa/api/review/domain/entity/Review.java
@@ -1,11 +1,0 @@
-package com.moa.moa.api.review.domain.entity;
-
-import com.moa.moa.api.member.domain.entity.Member;
-import com.moa.moa.api.shop.domain.entity.Shop;
-
-public class Review {
-    private Shop shop;
-    private Member member;
-    private Double score;
-    private String content;
-}
