+++ src/main/java/com/moa/moa/api/shop/naverreview/domain/entity/NaverReview.java
@@ -1,0 +1,9 @@
+package com.moa.moa.api.shop.naverreview.domain.entity;
+
+import com.moa.moa.api.shop.shop.domain.entity.Shop;
+
+public class NaverReview {
+    private Shop shop;
+    private Double avgScore;
+    private Long totalReview;
+}
