+++ src/main/java/com/moa/moa/api/shop/placeshop/domain/entity/placeShop.java
@@ -1,0 +1,9 @@
+package com.moa.moa.api.shop.placeshop.domain.entity;
+
+import com.moa.moa.api.place.place.domain.entity.Place;
+import com.moa.moa.api.shop.shop.domain.entity.Shop;
+
+public class placeShop {
+    private Place place;
+    private Shop shop;
+}
