+++ src/main/java/com/moa/moa/api/shop/wish/domain/entity/Wish.java
@@ -1,0 +1,9 @@
+package com.moa.moa.api.shop.wish.domain.entity;
+
+import com.moa.moa.api.member.member.domain.entity.Member;
+import com.moa.moa.api.shop.shop.domain.entity.Shop;
+
+public class Wish {
+    private Member member;
+    private Shop shop;
+}
