--- src/main/java/com/moa/moa/api/slope/domain/entity/Slope.java
@@ -1,9 +1,0 @@
-package com.moa.moa.api.slope.domain.entity;
-
-import com.moa.moa.api.place.domain.entity.Place;
-
-public class Slope {
-    private Place place;
-    private String name;
-    private String level;
-}
