--- src/main/java/com/moa/moa/api/wish/domain/entity/Wish.java
@@ -1,9 +1,0 @@
-package com.moa.moa.api.wish.domain.entity;
-
-import com.moa.moa.api.member.domain.entity.Member;
-import com.moa.moa.api.shop.domain.entity.Shop;
-
-public class Wish {
-    private Member member;
-    private Shop shop;
-}
